[iOS] Complete the minimum implementation for the package verb.
Not fully ocmpleted, but the verb now create a fully working app that
can be executed in a simulator or a device to run the tests.

Command calls restore and msbuild. We should move this to its own task
classes like in xamarin-macios, but it is a good first implementation
for a POC.