[iOS] Provide the basic xUnit iOS runner.
Provide the basic iOS runner for possible clients. The API provides an
entry point class that will perform all the required initialization
expected by xharness and will run the tests.

Clients must inherit from ApplicationEntryPoint and implement the
following properties/methods. Later call RunAsync, for example, in
ViewDidLoad.

*protected abstract IDevice Device { get; }*

Must return an object that implements the interface. The following code
is an example of the implementation using Xamarin.iOS.

```csharp
public class XamarinIOSExample : IDevice
{
  UIDevice device = UIDevice.CurrentDevice;

  public string BundleIdentifier => NSBundle.MainBundle.BundleIdentifier;

  public string UniqueIdentifier {
  {
    get {
    IntPtr handle = UIDevice.CurrentDevice.Handle;
    if (UIDevice.CurrentDevice.RespondsToSelector (new Selector ("uniqueIdentifier")))
      return NSString.FromHandle (objc_msgSend (handle, Selector.GetHandle("uniqueIdentifier")));
    return "unknown";
  }

  public string Name => device.Name;
  public string Model => device.Model;
  public string SystemName => device.SystemName;
  public string SystemVersion => device.SystemVersion;
  public string Locale => NSLocale.CurrentLocale.Identifier;
}

```

*protected abstract TestRunner TestRunner { get; }*

Simply state the type of runner to use.

*protected abstract string IgnoreFilesDirectory { get; }*

Return the location in which traits and ignores have been added. The
patterns are as follow:

* ignore files: Files with the pattern *.ignore
* trait files: There are two available file names:
  - xunit-excludes.txt: For the traits to be ignored by the xunit
  runner.
  - nunit-excludes.txt: For the categories to be ignored inthe nunit
  runner.

*protected abstract void TerminateWithSuccess();*

Should terminate the app with a valid result. Example:

```csharp
void InternalTerminateWithSuccess ()
{
  // For WatchOS we're terminating the extension, not the watchos app itself.
  Console.WriteLine ("Exiting test run with success");
  Selector s = new Selector ("terminateWithSuccess");
  UIApplication.SharedApplication.PerformSelector (s, UIApplication.SharedApplication, 0);
}

public override void TerminateWithSuccess()
{
  BeginInvokeOnMainThread (TerminateWithSuccess);
}

```

Notes: This API is not set in stone.